chore: rename `OrderProgressBarV2` to `OrderProgressBar`
