+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/constants.ts
