+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/container/ProgressImageWrapper.tsx
