+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/container/RenderProgressTopSection.tsx
