+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/container/StepComponent.tsx
