+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/container/StepsWrapper.tsx
