+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/container/TopSections.tsx
