+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/container/styled.ts
