+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/helpers.ts
