+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/sharedStyled.ts
