+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/steps/AnimatedToken.tsx
