+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/steps/FinishedStep.tsx
