+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/steps/InitialStep.tsx
