+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/steps/SolvingStep.tsx
