+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/steps/styled.ts
