+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/styled.ts
