+++ apps/cowswap-frontend/src/common/pure/OrderProgressBar/types.ts
